+++ docs/client-apis/README.md
@@ -2,9 +2,9 @@
 
 Low-level client APIs, independent of specific UI screens or flows:
 
-* [Screens, Layouts, and Nav](screens-layouts-nav.md)
-* [App Context](app-context.md)
-* [Logging](logging.md)
-* [Actions](actions.md)
-* [User Messaging](user-messaging.md)
-* [Loadable](loadable.md)
+- [Screens, Layouts, and Nav](screens-layouts-nav.md)
+- [App Context](app-context.md)
+- [Logging](logging.md)
+- [Actions](actions.md)
+- [Status](status.md)
+- [Loadable](loadable.md)
