+++ templates/faves/admin/app/AuthConfig.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import firebase from 'firebase';
 import {Account} from '@toolkit/core/api/Auth';
 import {useApi} from '@toolkit/core/api/DataApi';
-import {useUserMessaging} from '@toolkit/core/client/UserMessaging';
+import {useUserMessaging} from '@toolkit/core/client/Status';
 import {FirebaseAuthService} from '@toolkit/providers/firebase/client/AuthService';
 import {UnauthorizedError} from '@toolkit/tbd/CommonErrors';
 import {GET_USER} from '@app/common/Api';
