+++ templates/faves/admin/app/screens/SendNotificationModal.tsx
@@ -10,7 +10,7 @@ import {StyleSheet, View} from 'react-native';
 import {Button, TextInput, Title, useTheme} from 'react-native-paper';
 import {useApi} from '@toolkit/core/api/DataApi';
 import {User} from '@toolkit/core/api/User';
-import {useUserMessaging} from '@toolkit/core/client/UserMessaging';
+import {useUserMessaging} from '@toolkit/core/client/Status';
 import {useNav} from '@toolkit/ui/screen/Nav';
 import {Screen} from '@toolkit/ui/screen/Screen';
 import {SEND_ADMIN_NOTIF} from '@app/common/Api';
